Default an interface to a concrete type
When deserializing, if we encounter a property that returns collection
interface, have the deserializer default to creating a concrete type of
that interface.  This assumes that there is a concrete type whose name
corresponds to the interface name.  We do this since in WinRT we cannot
expose concerete interface collections fom the public API